Move accessibility validation test to PaparazziPluginTest
