+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/validate-accessibility/build.gradle
@@ -1,0 +1,27 @@
+plugins {
+  id 'com.android.library'
+  id 'kotlin-android'
+  id 'app.cash.paparazzi'
+}
+
+repositories {
+  maven {
+    url "file://${projectDir.absolutePath}/../../../../../build/localMaven"
+  }
+  mavenCentral()
+  google()
+}
+
+android {
+  namespace 'app.cash.paparazzi.plugin.test'
+  compileSdk libs.versions.compileSdk.get() as int
+  defaultConfig {
+    minSdk libs.versions.minSdk.get() as int
+  }
+}
+
+tasks.withType(Test).configureEach {
+  testLogging {
+    showStandardStreams true
+  }
+}
