--- paparazzi/paparazzi/src/main/java/app/cash/paparazzi/accessibility/AccessibilityCheckConfig.kt
@@ -1,8 +1,0 @@
-package app.cash.paparazzi.accessibility
-
-/**
- * Configuration options for [Paparazzi.validateLayoutAccessibility]
- */
-data class AccessibilityCheckConfig(
-  val shouldValidate: Boolean = false
-)
